leaderboard_client/requirements.txt
flask
psycopg2
Flask-SQLAlchemy
python-dotenv
flask-cors